added module to setup the ipcMain handlers. #136.
